+++ src/main/java/com/playtomic/tests/wallet/respository/entity/WalletEntity.java
@@ -1,5 +1,7 @@
 package com.playtomic.tests.wallet.respository.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -21,6 +23,6 @@ public class WalletEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private long amount;
+    private BigDecimal amount;
 
 }
